refactor: add useLoginMutation hook
