+++ src/hooks/auth.ts
@@ -1,46 +1,16 @@
-import { useRouter } from 'next/navigation';
-import { useMutation } from '@tanstack/react-query';
 import { useSetAtom } from 'jotai';
 
 import { fetchEmailLogin, fetchNaverLogin } from '@/fetch/auth';
 import { authTokenManager, isLoggedInAtom } from '@/store/auth';
 import { EmailLoginType } from '@/types/join';
-import { LoginResponse } from '@/types/response';
+import { useLoginMutation } from './useAuthMutation';
 
 export function useEmailLogin(onErrorHandler: (error: string) => void) {
-  const router = useRouter();
-  const setIsLoggedIn = useSetAtom(isLoggedInAtom);
-
-  return useMutation({
-    mutationFn: (data: EmailLoginType) => fetchEmailLogin(data),
-    onSuccess: ({ data }: LoginResponse) => {
-      authTokenManager.setTokens(data.accessToken, data.refreshToken);
-      setIsLoggedIn(true);
-      router.push('/');
-    },
-    onError: (error) => {
-      authTokenManager.clearTokens();
-      setIsLoggedIn(false);
-      onErrorHandler(error.message);
-    },
-  });
+  return useLoginMutation<EmailLoginType>(fetchEmailLogin, onErrorHandler);
 }
 
-export function useNaverLogin(onErrorHandler: (error: string) => void) {
-  const setIsLoggedIn = useSetAtom(isLoggedInAtom);
-
-  return useMutation({
-    mutationFn: fetchNaverLogin,
-    onSuccess: ({ data }: LoginResponse) => {
-      authTokenManager.setTokens(data.accessToken, data.refreshToken);
-      setIsLoggedIn(true);
-    },
-    onError: (error) => {
-      authTokenManager.clearTokens();
-      setIsLoggedIn(false);
-      onErrorHandler(error.message);
-    },
-  });
+export function useNaverLogin() {
+  return useLoginMutation(fetchNaverLogin);
 }
 
 export function useLogout() {
